### 2023-01-10 #5061 @editorialbot
[PRE REVIEW]: Brainchop: In-browser MRI volumetric segmentation and rendering
**Submitting author:** @Mmasoud1 (<a href="http://orcid.org/0000-0002-5365-242X">Mohamed Masoud</a>)
**Repository:** https://github.com/neuroneural/brainchop
**Branch with paper.md** (empty if default branch):
**Version:** v1.4.0
