+++ paper/paper.bib
@@ -22,7 +22,7 @@
  	title = {TensorFlow.Js: Machine Learning for the Web and Beyond},
 	journal = {arXiv},
 	archiveprefix = {arXiv},
-	doi = {10.48550/arXiv:1901.05350v2},
+	doi = {10.48550/arXiv.1901.05350},
 	year = 2019,
 	month = Feb
 }
