+++ README.md
@@ -70,19 +70,18 @@ To see the tool in action please click  [here](https://neuroneural.github.io/bra
 
 ## News!
 
-Google, Tensorflow community spotlight award for brainchop on [Linkedin](https://www.linkedin.com/posts/tensorflow-community_github-neuroneuralbrainchop-brainchop-activity-6978796859532181504-cfCW?utm_source=share&utm_medium=member_desktop) and [Twitter](https://twitter.com/TensorFlow/status/1572980019999264774)
+* Google, Tensorflow community spotlight award for brainchop (Sept 2022) on [Linkedin](https://www.linkedin.com/posts/tensorflow-community_github-neuroneuralbrainchop-brainchop-activity-6978796859532181504-cfCW?utm_source=share&utm_medium=member_desktop) and [Twitter](https://twitter.com/TensorFlow/status/1572980019999264774)
 
 <div align="center">
 
 ![Interface](https://github.com/neuroneural/brainchop/blob/master/css/images/TF_CommunityAward.png)
 
-
-**Google, TensorFlow Community Spotlight Award for Brainchop project (Sept 2022)**
 </div>
 
 <br>
+<br>
 
-Brainchop  invited to [Pytorch](https://pytorchconference22.splashthat.com/) flag conference, New Orleans, Louisiana (Dec 2022) 
+* Brainchop  invited to [Pytorch](https://pytorchconference22.splashthat.com/) flag conference, New Orleans, Louisiana (Dec 2022) 
 
 <div align="center">
 
