### 2023-01-23 #5061 @richford
> 👋 @62442katieb & @richford & @mwegrzyn, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I'd be happy to as well.
