+++ test/spec.js
@@ -1,0 +1,2 @@
+var expect = chai.expect;
+
