+++ test/bc_test/mainMeshNetFunctions_test.js
@@ -1,0 +1,301 @@
+/*
+=========================================================
+* Brainchop - v1.4.0 - ((------Testing-------))
+=========================================================
+* Discription:  A user interface for whole brain segmentation
+*               Input shape : [1, D, H, W, 1] e.g. [1, 256, 256, 256, 1]                
+*               Model : Meshnet or similar     
+*
+* Authors:  Mohamed Masoud  and Sergey Plis  - 2023
+=========================================================
+=========================================================
+         Main MeshNet Functions Unit Test (Mocha)
+=========================================================*/ 
+
+
+describe("Main Brainchop Functions", function () {
+  describe('#generateColors()', function () {
+    it('return  colors array of string', function () {
+       expect(generateColors(100, 50,  3)).to.eql([ "hsla(0,100%,50%)", "hsla(120,100%,50%)", "hsla(240,100%,50%)" ]);
+    });
+  });
+
+  describe('#getRgbObject()', function () {
+    it('return RGB string as object', function () {
+       expect(getRgbObject( "rgb(255,0,0)" )).to.eql({ r: 255, g: 0, b: 0 });
+    });
+  });
+
+  describe('#hslToRgb()', function () {
+    it('return HSL conversion to RGB as string', function () {
+       expect(hslToRgb( "hsla(0,100%,50%)" )).to.equal("rgb(255,0,0)");
+    });
+  });
+
+  describe('#rgbToHex()', function () {
+    it('return RGB conversion to HEX as string', function () {
+       expect(rgbToHex( { r: 255, g: 0, b: 0 } )).to.equal("#ff0000");
+    });
+  });
+
+  describe('#intersect()', function () {
+    it('return intersection 1D array ', function () {
+       expect(intersect([0,1,1], [0,2,2])).to.eql([0]);
+    });
+  });
+
+  describe('#diceCoefficient()', function () {
+    it('return dice Coefficient number', function () {
+       expect(diceCoefficient([0,1,1], [0,2,2]).toFixed(1)).to.equal('0.3');
+    });
+  });
+
+  describe('#intersect()', function () {
+    it('return intersection 1D array ', function () {
+       expect(intersect([0,1,1], [0,2,2])).to.eql([0]);
+    });
+  });
+
+  describe('#getMaxRegionMaskByContour()', function () {
+    it('return  contour pixels value as Uint8Array ', function () {
+       expect(getMaxRegionMaskByContour({ width: 1, height: 1, data:[0, 0, 0, 0]}) ).to.be.an('Uint8Array');
+    });
+  });  
+
+  describe('#postProcessSlices3D()', function () {
+    it('return 2D labels array outputSlices after filtering noisy 3d regions ', function () {
+       expect(postProcessSlices3D( [ [0,0,0,0,  0,1,1,0,  0,0,0,0],
+                       [0,0,0,0,  0,0,1,1,  0,0,0,0],
+                       [0,0,0,0,  0,0,0,1,  0,1,1,0] ], 3, 4)).to.eql([ [0,0,0,0,  0,1,1,0,  0,0,0,0],
+                                                                        [0,0,0,0,  0,0,1,1,  0,0,0,0],
+                                                                        [0,0,0,0,  0,0,0,1,  0,0,0,0] 
+                                                                      ]);
+    });
+  });
+
+  describe('#randn_bm()', function () {
+    it('return Standard Normal variate using Box-Muller transformation ', function () {
+       expect(randn_bm()).to.be.a('number') ;
+    });
+  });
+
+
+  describe('#checkInside()', function () {
+    it('return true or false  ', function () {
[…]
