+++ test/README.md
@@ -1,0 +1,19 @@
+
+# Brainchop  [![Version](https://img.shields.io/badge/Version-1.4.0-brightgreen)]() [![JS ](https://img.shields.io/badge/Types-JavaScript-blue)]()  [![MIT-License ](https://img.shields.io/badge/license-MIT-green)](https://github.com/neuroneural/brainchop/blob/master/LICENSE) [![Pass ](https://img.shields.io/badge/Pass-OK-green)](https://neuroneural.github.io/brainchop/test/runner.html)
+
+
+
+
+## Unit Testing
+
+**Mocha** unit testing is available in the browser with this [link](https://neuroneural.github.io/brainchop/test/runner.html).
+
+
+<br />
+<br />
+<div align="center">
+
+<img src='https://github.com/neuroneural/brainchop/blob/master/css/logo/TReNDS_logo.jpg' width='300' height='100'></img>
+
+**Mohamed Masoud - Sergey Plis - 2023**
+</div>
