+++ test/spec.js
@@ -1,2 +1,0 @@
-var expect = chai.expect;
-
