### 2023-01-24 #5098 @samhforbes
Hi @mwegrzyn, @62442katieb, @richford thanks for agreeing to review. This is our review thread. Please generate your checklist with the above, and raise any issues about the software in the issues tab of the Brainchop repository directly, linking back to here, Any wider issues you'd like to discuss please flag here of course, and let me know if there's anything I can help with. Thanks again!
