### 2023-02-28 neuroneural/brainchop#17 @richford
[Documentation]: Add to acknowledgments
In the "About" model on brainchop.org, you mention:

> **Special thanks** to Kevin Wang and Alex Fedorov for discussions and pre-trained Meshnet models.

Would it be appropriate to include them in the [acknowledgments section of paper.md](https://github.com/neuroneural/brainchop/blob/bc1fa4ba4d4b99eca9191ec9e5abb9370c649138/paper/paper.md?plain=1#L88) as well?

ref: [JOSS Review](https://github.com/openjournals/joss-reviews/issues/5098)
