### 2023-02-28 #5098 @62442katieb
Went through the paper and checklist and I have a few notes:
The docs have a lot of information, with snippets of example code throughout, and the README + paper have a nice schematic of how Brainchop works, but it might be helpful to have a start-to-finish example solving a real-world problem.

The paper mentions that being browser-based avoids privacy issues. As it seems the authors intend this tool to be used by clinicians, it would be helpful to more explicitly spell out in the paper and docs _how_ this avoids privacy issues (e.g., nothing is saved to an external server, the tool can be installed and run locally, etc.) as not all clinicians are familiar with the implications of browser-based vs. server-side processing.

I'm testing out Brainchop in Firefox (v109.0.1) on an M1 MacBook Pro and many of the segmentation options won't run, with error messages such as: `This model needs texture size of minimum 16384, while current browser supports only 8192` and `This option needs a dedicated graphics card`. The system requirements are noted in the wiki, but it would be helpful to present those limitations more clearly in the paper and/or by denoting which segmentation options have specific system requirements on the browser client. Some options suggest I should use Safari instead, which would be helpful information to have up front (esp. as Figure 1 suggests Chrome and Firefox are the browsers of choice). As the authors intend this tool to increase accessibility, making these requirements/limitations more clear in the tool and in the paper would be helpful for users, especially those with limited access to computational resources.

The paper would benefit from explicitly comparing Brainchop to existing automated segmentation tools, in addressing the state of the field. The authors make it clear that the main benefit of this tool over others is its lack of installation requirements and ease of use, but that point would be more compelling if the authors compared it to the installation/usability of the tools it aims to supersede.

Very cool tool, overall! I think a few tweaks and it's good to go!
