### 2023-02-28 neuroneural/brainchop#18 @richford
[Bug]: Sidebar does not scroll on home page
On the home page of brainchop.org, the sidebar (the one with options "Open Brain T1 MRI," "Segmentation Options," etc) does not scroll and is cut off at the bottom on my Macbook. Would it be possible to make the sidebar scrollable?

Also, perhaps you could make each of the sidebar cards collapsible so that the user could hide them when they don't want them.

ref: [JOSS Review](https://github.com/openjournals/joss-reviews/issues/5098)
