### 2023-02-28 neuroneural/brainchop#19 @richford
[Bug]: Segmentation model info popup cuts off content
The info popup that comes up when you click on the <img width="27" alt="Screenshot 2023-02-28 at 11 52 45 AM" src="https://user-images.githubusercontent.com/7978135/221963732-0f0646e7-6f04-43dd-a937-3481befa84d1.png">
 button next to the segmentation model selection dropdown has a couple issues:

- It cuts off the content when the information text is over a certain length. For me, this happened for the "FS aparc+aseg Atlas 104 (failsafe)" model.
- If I click the info button for one model and then click on the info popup, the popup box stays open. If I then select another model, sometimes the popup content does not change and other times it presents information that is not that same as when I initially click the info button for that model, instead saying something like "This model needs dedicated graphics card" along with a checkbox that doesn't do anything.

ref: [JOSS Review](https://github.com/openjournals/joss-reviews/issues/5098)
