### 2023-02-28 #5098 @richford
@emdupre, I went through the paper and source code. Overall, it is very impressive work. I opened a few issues on the source repository related to UX and the paper itself. They are also linked in this thread.

I cannot verify the performance claims in the documentation because I don't have that exact hardware. But they are very plausible given the performance I saw on my M1 MacBook.

@62442katieb also brought up some good point that I'd like to echo. Namely,
- The paper should say a bit more about the "state of the field," even if that means discussing "offline" segmentation tools.
- The documentation would benefit from a start to finish example (including preprocessing) using a NIfTI file from an openly available repository.

I enjoyed reading the paper and test driving the tool. With a few minor modifications, I will give this submission an enthusiastic approval!
