+++ README.md
@@ -69,10 +69,19 @@ To see the tool in action please click  [here](https://neuroneural.github.io/bra
 
 ## News!
 
+* Brainchop 1-page abstract and poster is accepted in 20th IEEE International Symposium on Biomedical Imaging ([ISBI 2023](https://2023.biomedicalimaging.org/en/))
+
+<div align="center">
+   <img src="https://github.com/neuroneural/brainchop/blob/master/css/news/ISBI_2023.png"  width="60%">
+</div>
+
+<br>
+<br>
+
 * Google, Tensorflow community spotlight award for brainchop (Sept 2022) on [Linkedin](https://www.linkedin.com/posts/tensorflow-community_github-neuroneuralbrainchop-brainchop-activity-6978796859532181504-cfCW?utm_source=share&utm_medium=member_desktop) and [Twitter](https://twitter.com/TensorFlow/status/1572980019999264774)
 
 <div align="center">
-   <img src="https://github.com/neuroneural/brainchop/blob/master/css/images/TF_CommunityAward.png"  width="60%">
+   <img src="https://github.com/neuroneural/brainchop/blob/master/css/news/TF_CommunityAward.png"  width="60%">
 </div>
 
 <br>
@@ -81,7 +90,7 @@ To see the tool in action please click  [here](https://neuroneural.github.io/bra
 * Brainchop  invited to [Pytorch](https://pytorchconference22.splashthat.com/) flag conference, New Orleans, Louisiana (Dec 2022) 
 
 <div align="center">
-  <img src="https://github.com/neuroneural/brainchop/blob/master/css/images/Pytorch_Poster.jpg"  width="50%">
+  <img src="https://github.com/neuroneural/brainchop/blob/master/css/news/Pytorch_Poster.jpg"  width="50%">
 </div>
 
 
