### 2023-03-02 neuroneural/brainchop#19 @Mmasoud1
Thank you, I pushed fixes for the bug, basically we have two popups, Info (User click)  and Warning(Auto popup). The checkbox is for warning popup as "Don't show again".  Please check for the fixes.
