### 2023-03-03 neuroneural/brainchop#18 @Mmasoud1
Thank you,  I added a scroll bar to the left mini forms of tool options. Try, please.

For the collapsible enhancement,  currently, it raises some styling issues with the UI while try embedding the mini forms with accordion control since Webix doesn't support such an integration. We will keep that enhancement idea for future versions subject to finding a feasible workaround or performing major stylish refactoring to the side forms.  Till then we may consider the current issue closed for now.
