### 2023-03-06 #5098 @mwegrzyn
Dear [@Mmasoud1](https://github.com/Mmasoud1),

thank you for your work on brainchop. The following are some comments I have regarding the package and the manuscript:

#### use of brainchop website
- [ ] the website is well-designed. However, depending on screen size, not all elements are visible. For example, only part of the papaya viewers are visible on a smaller screen
- [ ] regarding hardware and software requirements, the manuscript states that the tool might be particularly beneficial in rural areas and developing countries. However, it seems to me that the tool demands a lot from the user’s machine. In fact, I was not able to use it on various desktop and laptop computers (both Linux and Windows), and only could use it partly running Chrome (but not Firefox). Many of the models were inaccessible to me (“This model may need dedicated graphics card”). Performing “Subvolume GWM (failsafe)” took minutes to complete. Due to this, I did not check the “Performance: If there are any performance claims of the software, have they been confirmed?” part of the JOSS checklist, as I think that the requirements of the package should be communicated in a more explicit way

#### validation of brain extraction tools
- [ ] I was wondering how well the different brain extraction algorithms perform and how they were pre-trained. More information on this would be helpful. In the acknowledgements I read “The authors would like to thank Kevin Wang and Alex Fedorov for discussions and pre-trained Meshnet models.”. Could you elaborate further? Thank you
- [ ] I was not able to use all Segmentation Options of the tool, due to my insufficient hardware and/or software setup, but when using “Extract the Brain (FAST)” on one of my own images (https://openneuro.org/datasets/ds001419/versions/1.0.1/file-display/sub-01:anat:sub-01_T1w.nii.gz) I was not satisfied with the result. It contains both false-positives (a substantial part of the skull labelled as brain) and false-negatives (a substantial part of orbitofrontal brain areas not labelled as brain). This is shown below. I am used to imperfect brain extraction, for example from the FSL brain extraction tools, but there I can manually tweak the parameters to get better results. How can these issues be addressed in brainchop? Can I change certain settings to improve results?
![FP-skull](https://user-images.githubusercontent.com/10476281/223171639-ab81880c-58fa-4c81-ad54-79e5abf7f7d5.png)
![FN-frontal](https://user-images.githubusercontent.com/10476281/223171668-538d258f-0761-4023-8fa4-1746edd20584.png)

- [ ] How will the pre-trained models perform for brains with atrophy, lesions or for developing brains?

#### manuscript
- [ ] I found the manuscript well-written and helpful. The motivation is clear to me. I would like to know however, how the tool compares to other software packages, for example MricroGL, SPM, FSL, FreeSurfer, ANTs, and/or fMRIprep pipelines). Due to this, I did not check the “State of the field: Do the authors describe how this software compares to other commonly-used packages?” part of the JOSS checklist
- [ ] line 58: “Papaya (Papaya dev. team, 2019) viewers is used” → this should read “Papaya (Papaya dev. team, 2019) viewer is used”
- [ ] line 61: “Three.js (Cabello & et, 2010) ” → this should read “Three.js (Cabello et al., 2010) ” and in the references, “& et, al.” should be corrected

#### documentation in wiki
- [ ] contribution guidelines and ways to get help are well-explained. However, I was not able to find much information on the different Models for the Segmentation Options. Due to this, I did not check the “Functionality documentation: Is the core functionality of the software documented to a satisfactory level” box on the JOSS checklist. Could you elaborate on the different models and the function to “Load tfjs Models” (see screenshot below)?
![models](https://user-images.githubusercontent.com/10476281/223171409-d626fec8-5d9a-43b7-82fd-08fda80c2a7e.png)
- [ ] The mocha testing does show an error for #normalizeVolumeData(): https://neuroneural.github.io/brainchop/test/runner.html

I hope you find these comments helpful. Please let me know if you have any questions regarding the points I raised. Thank you for your time and consideration.

Best,
Martin Wegrzyn
