### 2023-03-07 #5098 @mwegrzyn
Dear [@Mmasoud1](https://github.com/Mmasoud1),

thank you for your message. I am looking forward to you revision. I followed the link in the wiki to https://neuroneural.github.io/brainchop/test/runner.html and get the following message (on Debian with Firefox and on Windows with Firefox/Chrome/Edge):

![Screenshot from 2023-03-07 08-04-26](https://user-images.githubusercontent.com/10476281/223350622-1fb61d87-b1e4-46da-b37e-7fb2bfc5cbf2.png)

I hope this helps.

Best,
Martin Wegrzyn
