+++ paper/paper.bib
@@ -68,7 +68,7 @@
 }
 
 @misc{threejs,
-  author = {{Cabello}, R. and {et}, al.}, 
+  author = {{Cabello}, R. and {et al}.}, 
 	title = {Three.js},
 	publisher = {GitHub},
 	journal = {GitHub repository},
@@ -87,5 +87,32 @@
   url     = {https://onlinelibrary.wiley.com/doi/abs/10.1087/20150211},
   year    = {2015}
 }
+
+@article{Friston:1994,
+  author  = {{Friston}, KJ. and {Holmes}, AP. and {Worsley}, KJ. and {Poline}, JB. and {Frith}, CD. and {Frackowiak}, RSJ.},
+  title   = {Statistical parametric maps in functional imaging: a general linear approach},
+  journal = {Human Brain Mapping},
+  volume  = {2},
+  number  = {4},
+  pages   = {189-210},
+  doi     = {10.1002/hbm.460020402},
+  url     = {https://onlinelibrary.wiley.com/doi/10.1002/hbm.460020402},
+  year    = {1994}
+}
+
+@article{FS:1999,
+  author  = {{Dale}, AM. and {Fischl}, B. and {Sereno}, MI.},
+  title   = {Cortical surface-based analysis: I. Segmentation and surface reconstruction},
+  journal = {Neuroimage},
+  volume  = {9},
+	number  = {2},
+	pages   = {179-194},
+	year    = {1999},
+	issn    = {1053-8119},
+	doi     = {https://doi.org/10.1006/nimg.1998.0395},
+	url     = {https://www.sciencedirect.com/science/article/pii/S1053811998903950}
+}
+
+
 
 
