+++ paper/paper.bib
@@ -18,7 +18,7 @@
 }
 
 @article{tensorflow-js,
-	author = {{Smilkov}, D. and {Thorat}, N. and {Assogba}, Y. and {Yuan}, A. and {Kreeger}, N. and {et}, al.},
+	author = {{Smilkov}, D. and {Thorat}, N. and {Assogba}, Y. and {Yuan}, A. and {Kreeger}, N. et al.},
  	title = {TensorFlow.Js: Machine Learning for the Web and Beyond},
 	journal = {arXiv},
 	archiveprefix = {arXiv},
@@ -68,7 +68,7 @@
 }
 
 @misc{threejs,
-  author = {{Cabello}, R. and {et al}.}, 
+  author = {{Cabello}, R. et al.}, 
 	title = {Three.js},
 	publisher = {GitHub},
 	journal = {GitHub repository},
