+++ paper/paper.bib
@@ -67,7 +67,7 @@
 }
 
 @misc{threejs,
-  author = {{Cabello}, R. and {Herzog}, M. and {others}}, 
+  author = {{Cabello}, R. and {Qualia}, A. and {others}}, 
 	title = {Three.js},
 	publisher = {GitHub},
 	journal = {GitHub repository},
