### 2023-03-08 #5098 @Mmasoud1
Dear @62442katieb , thank you very much  for  your time and valuable comments.

**Example usage**:  We added a [show case](https://github.com/neuroneural/brainchop/wiki/Show-Case
) example  to the doc.

**Data privacy statement**:  In the _statement of needs_ section, we added the suggested phrase in [line 28]( https://github.com/neuroneural/brainchop/blob/master/paper/paper.pdf)

**Segmentation models** :   Despite the high diversity of H/W and S/W resources that exist on the users' side, we did our best to give them feasible recommendations. In addition to the warning messages for the need for a dedicated graphics card, we also added a browser resources window    ![Screenshot 2023-03-08 at 05-52-15 Brainchop](https://user-images.githubusercontent.com/36082112/223694864-5683ebf8-a390-42cd-ac55-ac857fddd9c1.png)    showing end-user browser resources and a [link](https://docs.google.com/spreadsheets/d/1PU4p6oN4aBfTbXZ8pO6n8fc890Y5Ck0agsIZbm_FO3I/edit#gid=0) to verified S/W and H/W for each model.

**State of the field**: In the _statement of needs_ section, we added the state of the field paragraph in [lines 29-34.]( https://github.com/neuroneural/brainchop/blob/master/paper/paper.pdf)

Best,
