### 2023-03-08 #5098 @Mmasoud1
Dear @richford, thank you  for  your time, valuable comments and opening bug issues.

**Add author contributions** :     Contribution section added [lines 74-78]( https://github.com/neuroneural/brainchop/blob/master/paper/paper.pdf).

**Add to acknowledgments** :  Acknowledgments section modified.

**Sidebar does not scroll on home page [issue 18](https://github.com/neuroneural/brainchop/issues/18)** :   We added a scroll bar to the left mini forms of tool options  [e951fbd](https://github.com/neuroneural/brainchop/commit/e951fbd33d540fd131c32357482387d2275675bd)

https://user-images.githubusercontent.com/36082112/223699793-757f0f08-3151-4c44-826d-968c8037a9bd.mp4

**Segmentation model info popup cuts off content [issue 19](https://github.com/neuroneural/brainchop/issues/19)** : We pushed  Fixes for the bug [9f5cae7](https://github.com/neuroneural/brainchop/commit/9f5cae7e0e52dc8597906992ad3ad99974062ac0)

**State of the field**: In the _statement of needs_ section, we added the state of the field paragraph in [lines 29-34.]( https://github.com/neuroneural/brainchop/blob/master/paper/paper.pdf)

**Example usage**:  We added a [show case](https://github.com/neuroneural/brainchop/wiki/Show-Case
) example  to the doc.

Best,
