### 2023-03-08 #5098 @Mmasoud1
Dear @mwegrzyn,  thank you  for  your time, and valuable comments.

#### Use of brainchop website:
I tried to mediate this stylish conflict between Webix and Papaya  by adjusting the padding. This makes MRI planes visible for screen width in range > 1350  and subject also for more future improvements to fit smaller sizes.

#### Validation of brain extraction tools

I added two deeper Meshnet models ( 11 filters per layer ) for better brain extraction and masking accuracy : Extract the Brain (High Acc) and Compute Brain Mask (High Acc). Both models are tested with Apple M1 and  GeForce GTX 980.

https://user-images.githubusercontent.com/36082112/223703125-0e1d6d89-8a3f-4b2f-9cd3-b93c88cf7d78.mp4

For more info about Meshnet training please refer to this previous work of out lab [here](https://github.com/catalyst-team/neuro#example-segmentation-notebooks). We also plan to integrate training pipeline with brainchop repo for advanced users in version 3.0.0.

Although fast brain extraction model (i.e. Extract the Brain (FAST) ) has less accuracy than the aforementioned model, it is more suitable to run on machines with Internal or built-in graphics chip (e.g. HD, UHD) that normally unable to run 11 filters per layer.

#### Manuscript:
**State of the field**: In the _statement of needs_ section, we added the state of the field paragraph in [lines 29-34.]( https://github.com/neuroneural/brainchop/blob/master/paper/paper.pdf)

**Typos**:  Suggested typos corrections are applied  for Papaya and three.js. Thank you!.

#### Documentation in wiki:

**Load tfjs Models**:  Description section has been added to [wiki](https://github.com/neuroneural/brainchop/wiki/Inference-Model)

**Mocha test**: We fixed it  taking into consideration that  issue may arise due to slightly different floating point calculations between machines.

Best,
