### 2023-03-10 #5098 @Mmasoud1
> Thanks both for your reviews! @Mmasoud1 do update us here when these issues have been addressed! ......

Dear @samhforbes,  we addressed them to the best we can and closed all opening issues on brainchop repo.    if there is anything else still needed on our part please let me know.   Many thanks!
