+++ paper/paper.bib
@@ -17,13 +17,14 @@
 	month = Apr
 }
 
+
 @article{tensorflow-js,
-  author = {{Smilkov}, D. and {Thorat}, N. and {Assogba}, Y. and {Nicholson}, C. and {Kreeger}, N. and {Yu}, P. and {Cai}, S. and {Nielsen}, E. and {Soegel}, D. and {Bileschi}, S. and {others}},
-  title = {Tensorflow. js: Machine learning for the web and beyond},
-  journal = {Proceedings of Machine Learning and Systems},
-  volume = {1},
-  pages = {309-321},
-  year = {2019}
+  author = {{Smilkov}, D. and {Thorat}, N. and {Assogba}, Y. and {Nicholson}, C. and {Kreeger}, N. and {Yu}, P. and {Cai}, S. and {Nielsen}, E. and {Soegel}, D. and {others}},
+  title = {TensorFlow.Js: Machine Learning for the Web and Beyond},
+	journal = {arXiv},
+	archiveprefix = {arXiv},
+	doi = {10.48550/arXiv:1901.05350},
+	year = {2019}
 }
 
 @misc{NIfTI-Reader,
