+++ paper/paper.bib
@@ -3,8 +3,8 @@
   title = {End-to-end learning of brain tissue segmentation from imperfect labeling},
 	journal = {IEEE International Joint Conference on Neural Networks (IJCNN)},
 	doi = {10.1109/IJCNN.2017.7966333},
-	year = 2017,
-	month = May
+	year = {2017},
+	month = {May}
 }
 
 @article{Yu:2016,
@@ -13,8 +13,8 @@
 	journal = {arXiv},
 	archiveprefix = {arXiv},
 	doi = {10.48550/arXiv.1511.07122},
-	year = 2016,
-	month = Apr
+	year = {2016},
+	month = {Apr}
 }
 
 
@@ -29,7 +29,7 @@
 
 @misc{NIfTI-Reader,
 	title = {NIfTI Reader},
-	year = 2021,
+	year = {2017},
 	publisher = {GitHub},
 	journal = {GitHub repository},
 	url = {https://github.com/rii-mango/NIFTI-Reader-JS}
@@ -39,8 +39,8 @@
 @software{pyodide_2022,
   author       = {{Pyodide dev. team}},
   title        = {Pyodide},
-  month        = Apr,
-  year         = 2022,
+  month        = {Apr},
+  year         = {2022},
   publisher    = {Zenodo},
   version      = {0.20.0},
   doi          = {10.5281/zenodo.6430433},
@@ -53,16 +53,16 @@
 	title = {FastSurfer - A fast and accurate deep learning based neuroimaging pipeline},
 	journal = {Neuroimage},
 	doi = {10.1016/j.neuroimage.2020.117012},	
-	year = 2020,
-	month = Oct
+	year = {2020},
+	month = {Oct}
 }
 
 @misc{Papaya,
   author       = {{Papaya dev. team}},
 	title        = {Papaya},
 	publisher    = {GitHub},
-	month        = Mar,
-	year         = 2019,
+	month        = {Mar},
+	year         = {2019},
 	journal      = {GitHub repository},
 	url          = {https://github.com/rii-mango/Papaya}
 }
@@ -72,7 +72,7 @@
 	title = {Three.js},
 	publisher = {GitHub},
 	journal = {GitHub repository},
-	year = 2010,	
+	year = {2010},	
 	url = {https://github.com/mrdoob/three.js}
 }
 
