+++ paper/paper.bib
@@ -109,7 +109,7 @@
 	pages   = {179-194},
 	year    = {1999},
 	issn    = {1053-8119},
-	doi     = {https://doi.org/10.1006/nimg.1998.0395},
+	doi     = {10.1006/nimg.1998.0395},
 	url     = {https://www.sciencedirect.com/science/article/pii/S1053811998903950}
 }
 
