+++ paper/paper.md
@@ -15,9 +15,9 @@ authors:
       orcid: 0000-0003-0040-0365
       affiliation: "1, 2"
 affiliations:
-  - name: Tri-institutional Center for Translational Research in Neuroimaging and Data Science (TReNDS), Georgia State University, Georgia Institute of Technology, Emory University, Atlanta, GA, USA
+  - name: Tri-institutional Center for Translational Research in Neuroimaging and Data Science (TReNDS), Georgia State University, Georgia Institute of Technology, Emory University, Atlanta, United States of America
     index: 1
-  - name: Georgia State University, Department of Computer Science, Atlanta, GA, USA
+  - name: Department of Computer Science, Georgia State University, Atlanta, United States of America
     index: 2
 date: 9 December 2022
 bibliography: paper.bib
@@ -43,12 +43,12 @@ In order to deploy the PyTorch MeshNet model in the browser, there is a  need to
 
 # Preprocessing
 
-Brainchop is designed to support T1 weighted MRI volume segmentation. The input is read in Nifti format [@NIfTI-Reader]. T1 image needs to be in shape 256x256x256, scaled, and resampled to 1mm isotropic voxels as a preprocessing step for proper results. This preprocessing can be made in brainchop by using mri_convert.js which uses pyodide [@pyodide_2022] to deploy the conform function used by FastSurfer [@Henschel:2020] for reshaping, scaling, and re-sampling  MRI T1 raw image data as shown in  \autoref{fig:Convert-Enhance-Pipeline} (a).
+Brainchop is designed to support T1 weighted MRI volume segmentation. The input is read in Nifti format [@NIfTI-Reader]. T1 image needs to be in shape 256x256x256, scaled, and resampled to 1 mm isotropic voxels as a preprocessing step for proper results. This preprocessing can be made in brainchop by using mri_convert.js which uses Pyodide [@pyodide_2022] to deploy the conform function used by FastSurfer [@Henschel:2020] for reshaping, scaling, and re-sampling  MRI T1 raw image data as shown in  \autoref{fig:Convert-Enhance-Pipeline} (a).
 
 ![Brainchop preprocessing pipeline (a) Conform operation (b) MRI enhancement operations.\label{fig:Convert-Enhance-Pipeline}](ConvertAndEnhancePipeline.png)
 
 
-The rest of the preprocessing pipeline is to remove input noisy voxels and enhance input volume intensities to improve the segmentation accuracy \autoref{fig:Convert-Enhance-Pipeline} (b). In addition, brainchop also supports  MRI tissue cropping to speed up the inference process and lowering memory use. 
+The rest of the preprocessing pipeline is to remove input noisy voxels and enhance input volume intensities to improve the segmentation accuracy \autoref{fig:Convert-Enhance-Pipeline} (b). In addition, brainchop also supports  MRI tissue cropping to speed up the inference process and lower memory use. 
 
 
 
@@ -59,7 +59,7 @@ The advantage of MeshNet small size is due to its simple architecture and using 
 
 ![MeshNet architecture.\label{fig:MeshNet-Arch}](MeshNetArch.png)
 
-While MeshNet model has fewer parameters compared to the classical segmentation model U-Net, it also can achieve  a competitive DICE score as shown in \autoref{tab:Table-1}.
+While the MeshNet model has fewer parameters compared to the classical segmentation model U-Net, it also can achieve  a competitive DICE score as shown in \autoref{tab:Table-1}.
 
 
[…]
