### 2023-03-22 #5098 @samhforbes
Hi @Mmasoud1 thanks for flagging this - I'm not sure what's up with my browser, so I apologise.

Great! Looks like everyone is really pleased, and having had a little play with it, so am I.
Can you please check the version number is accurate, confirm the author list, and then archive it somewhere with a stable DOI (Zenodo, or figshare for example) - making sure the authors and title match that of the paper. Then could you please post the DOI here.
