### 2023-03-22 #5098 @Mmasoud1
Hi @samhforbes , thank you for the suggested changes, I reviewed the [version number](https://github.com/neuroneural/brainchop/releases/tag/v2.1.0), the author list and archived the latest version on [Zenodo](https://zenodo.org/record/7735848#.ZBtcYI7MJhH)  with same title and author list :

**Version:**     2.1.0
**DOI:**     10.5281/zenodo.7735848
