+++ paper/paper.bib
@@ -37,7 +37,7 @@
 
 
 @software{pyodide_2022,
-  author       = {{Pyodide dev. team}},
+  author       = {{Droettboom}, M. and {Chatham}, H. and {Yurchak}, R. and {Chua}, D. and {Choi}, G. and {Schreiner}, H. and {Abramowitz}, M. and {others}},
   title        = {Pyodide},
   month        = {Apr},
   year         = {2022},
@@ -60,11 +60,10 @@
 @misc{Papaya,
   author       = {{Papaya dev. team}},
 	title        = {Papaya},
-	publisher    = {GitHub},
+	publisher    = {GitHub Pages},
 	month        = {Mar},
 	year         = {2019},
-	journal      = {GitHub repository},
-	url          = {https://github.com/rii-mango/Papaya}
+	url          = {http://rii-mango.github.io/Papaya}
 }
 
 @misc{threejs,
