+++ README.md
@@ -1,5 +1,5 @@
 
-# Brainchop  [![Version](https://img.shields.io/badge/Version-1.4.0-brightgreen)]() [![JS ](https://img.shields.io/badge/Types-JavaScript-blue)]() [![HW ](https://img.shields.io/badge/HardWare-GPU-green)](https://github.com/neuroneural/brainchop/wiki/System-Requirements) [![MIT-License ](https://img.shields.io/badge/license-MIT-green)](https://github.com/neuroneural/brainchop/blob/master/LICENSE) [![tfjs](https://img.shields.io/badge/tfjs-Pre--trained%20Model-blue)](https://github.com/neuroneural/brainchop/tree/master/models/mnm_tfjs_me_test) 
+# Brainchop  [![Version](https://img.shields.io/badge/Version-2.1.0-brightgreen)]() [![JS ](https://img.shields.io/badge/Types-JavaScript-blue)]() [![HW ](https://img.shields.io/badge/HardWare-GPU-green)](https://github.com/neuroneural/brainchop/wiki/System-Requirements) [![MIT-License ](https://img.shields.io/badge/license-MIT-green)](https://github.com/neuroneural/brainchop/blob/master/LICENSE) [![tfjs](https://img.shields.io/badge/tfjs-Pre--trained%20Model-blue)](https://github.com/neuroneural/brainchop/tree/master/models/mnm_tfjs_me_test) 
 
 
 <div align="center">
@@ -114,7 +114,7 @@ To see the tool in action please click  [here](https://neuroneural.github.io/bra
 
 ## Citation
 
-Mohamed Masoud, Farfalla Hu and Sergey Plis. (2022). Brainchop : In-browser MRI volumetric segmentation and rendering (v1.4.0). Zenodo. https://doi.org/10.5281/zenodo.6565234
+Mohamed Masoud, Farfalla Hu and Sergey Plis. (2022). Brainchop : In-browser MRI volumetric segmentation and rendering (v2.1.0). Zenodo. https://doi.org/10.5281/zenodo.7735848
 
 ## Funding
 
